backend/requirements.txt
alembic==1.18.4
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
Mako==1.3.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomli==2.0.1
fastapi==0.110.0
uvicorn==0.29.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
psycopg2-binary
pydantic
python-multipart==0.0.9
aiosqlite==0.20.0
python-jose[cryptography]
pydantic[email]
passlib==1.7.4
bcrypt==3.2.2
boto3
aioboto3
